Fix number overflow issues with adjust operations on date

+++ legend-pure-m4/src/main/java/org/finos/legend/pure/m4/coreinstance/primitive/date/AbstractDateWithHour.java
@@ -37,7 +37,7 @@ abstract class AbstractDateWithHour extends AbstractDateWithDay implements DateT
     }
 
     @Override
-    public PureDate addHours(int hours)
+    public PureDate addHours(long hours)
     {
         if (hours == 0)
         {
