+++ legend-pure-m4/src/main/java/org/finos/legend/pure/m4/coreinstance/primitive/date/StrictDate.java
@@ -74,25 +74,25 @@ public final class StrictDate extends AbstractDateWithDay
     }
 
     @Override
-    public PureDate addHours(int hours)
+    public PureDate addHours(long hours)
     {
         throw new UnsupportedOperationException();
     }
 
     @Override
-    public PureDate addMinutes(int minutes)
+    public PureDate addMinutes(long minutes)
     {
         throw new UnsupportedOperationException();
     }
 
     @Override
-    public PureDate addSeconds(int seconds)
+    public PureDate addSeconds(long seconds)
     {
         throw new UnsupportedOperationException();
     }
 
     @Override
-    public PureDate addMilliseconds(int milliseconds)
+    public PureDate addMilliseconds(long milliseconds)
     {
         throw new UnsupportedOperationException();
     }
